[backend] close #71, users route ready. [database] update habit_category table.

+++ backend/services/userService.js
@@ -44,48 +44,39 @@ export async function findEmailPassword(email, password, callback) {
 }
 
 export async function findUserByEmail(email, callback) {
-  const [result] = await pool.query(`SELECT * FROM users WHERE email = ?`,
-    [email]);
+  const [result] = await pool.query(`SELECT * FROM users WHERE email = ?`, [email]);
   callback(result);
 }
 
-export async function findUserById(conn, id, callback) {
-  console.log(conn);
-  const result = await conn.query(`SELECT * FROM users WHERE id = ?`, [id]);
-  callback(result);
+export async function findUserById(id) {
+  const [result] = await pool.query('SELECT * FROM users WHERE id = ?', [id]);
+  return result;
 }
 
-export async function findUserByUsername(conn, profileName, callback) {
+export async function updateUser(userId, updatedFields) {
+  const fields = [];
+  const values = [];
+
+  for (const [key, value] of Object.entries(updatedFields)) {
+    fields.push(`${key} = ?`);
+    values.push(value);
+  }
+
+  const sql = `UPDATE users SET ${fields.join(', ')} WHERE id = ?`;
+  values.push(userId); // user ID at the end
+
+  const [result] = await pool.query(sql, values);
+  return result;
+}
+
+/*export async function findUserByUsername(conn, profileName, callback) {
   console.log(conn);
   const result = await conn.query(
     `SELECT * FROM users WHERE username LIKE ?`,
     ["%" + profileName + "%"]
   );
   callback(result);
-}
-
-export async function updateUser(conn, userId, updatedUser, callback) {
-  console.log(conn);
-  console.log("updated user:", updatedUser);
-  console.log("user id:", userId);
-  bcrypt.genSalt(10, function (err, salt) {
-    bcrypt.hash(updatedUser.password, salt, async function (err, hash) {
-      const result = await conn.query(
-        `UPDATE users
-    SET email = ?, password = ?, profileName = ?, profileImage = ?
-    WHERE id = ?`,
-        [
-          updatedUser.email,
-          hash,
-          updatedUser.profileName,
-          updatedUser.profileImage,
-          userId,
-        ]
-      );
-      callback(result);
-    });
-  });
-}
+}*/
 
 export async function updateProfileName(conn, userId, profileName, callback) {
   console.log(conn);
@@ -109,12 +100,7 @@ export async function updateProfileImage(conn, userId, profileImage, callback) {
   callback(result);
 }
 
-export async function deleteUser(conn, userId, callback) {
-  console.log(conn);
-  const result = await conn.query(
-    `DELETE FROM users
-    WHERE id = ?`,
-    [userId]
-  );
-  callback(result);
+export async function deleteUser(userId) {
+  const [result] = await pool.query('DELETE FROM users WHERE id = ?', [userId]);
+  return result;
 }
